cdk/pyproject.toml
[project]
name = "cdk"
version = "0.1.0"
description = "Pupper Dog Adoption Application CDK Stack"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib==2.201.0",
    "cdk-dynamo-table-view>=0.2.488",
    "constructs>=10.0.0,<11.0.0",
    "boto3>=1.34.0",
    "Pillow>=10.0.0",  # For image processing
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "moto>=4.2.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "cdk-nag>=2.0.0",
    "requests>=2.31.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=functions",
    "--cov=cdk",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--verbose"
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | cdk.out
)/
'''

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
